requirements.txt
ultralytics==8.0.136
streamlit==1.24.0
py-cpuinfo
numpy
firebase_admin
opencv-python
pillow
